[Chore] Constraint Extension 코드들 분리(#16)
